% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{basis,scNMFSet-method}
\alias{basis,scNMFSet-method}
\title{Basis matrix accessor}
\usage{
\S4method{basis}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing basis matrix}
}
\value{
List of basis matrices
}
\description{
Basis matrix accessor
}
