% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{basis<-,scNMFSet-method}
\alias{basis<-,scNMFSet-method}
\title{Basis matrix assignment}
\usage{
\S4method{basis}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing basis matrix}

\item{value}{New basis matrices to be substituted}
}
\value{
Object with basis matrix modified
}
\description{
Basis matrix assignment
}
