% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{basis<-}
\alias{basis<-}
\title{Modify basis matrices
 
Can be used to access and modify basis matrices}
\usage{
basis(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Basis matrix to be substituted}
}
\value{
Input object with updated basis matrices

Input object with updated basis matrices
}
\description{
Modify basis matrices
 
Can be used to access and modify basis matrices
}
\examples{
set.seed(1)
s <- scNMFSet(count=matrix(rpois(n=12,lambda=3),4,3))
s <- vb_factorize(s, ranks=3)
basis(s)[[1]] <- apply(basis(s)[[1]],1:2,round,digits=3)
basis(s)
}
