% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfindR.R
\docType{package}
\name{ccfindR}
\alias{ccfindR}
\alias{ccfindR-package}
\title{ccfindR: Cancer Clone FindeR}
\description{
This package contains tools and utilities for cell-type discovery
using single-cell transcriptomic data while evaluating significance
of the depth of clustering \insertCite{woo_etal}{ccfindR}.
}
\references{
\insertRef{woo_etal}{ccfindR}
}
