% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{cells,scNMFSet-method}
\alias{cells,scNMFSet-method}
\title{Cell annotation accessor}
\usage{
\S4method{cells}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing cell annotation}
}
\value{
Cell annotation data frame
}
\description{
Cell annotation accessor
}
