% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{cells<-,scNMFSet-method}
\alias{cells<-,scNMFSet-method}
\title{Cell annotation assignment}
\usage{
\S4method{cells}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing cell annotation}

\item{value}{New cell annotation to be substituted}
}
\value{
Object with modified cell annotation
}
\description{
Cell annotation assignment
}
