% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{cells<-}
\alias{cells<-}
\title{Modify cell annotation
 
Can be used to access and modify cell annotation}
\usage{
cells(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{cell annotation data frame to be substituted}
}
\value{
Input object with updated cell annotation

Input object with updated cell annotation
}
\description{
Modify cell annotation
 
Can be used to access and modify cell annotation
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
cells(s) <- letters[1:3]
cells(s)
}
