% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{cells}
\alias{cells}
\title{Cell annotation of Object}
\usage{
cells(object)
}
\arguments{
\item{object}{Object of class \code{scNMFSet}.}
}
\value{
Either \code{NULL} or data frame.'
}
\description{
Retrieve or set the cell annotation of object
}
\details{
\code{cells(object)} will return the cell annotation. 
\code{cells(object) <- value}
can be used to assign to the \code{cells} slot of \code{object}.
}
\examples{
mat <- matrix(rpois(n=12,lambda=3),4,3)
colnames(mat) <- LETTERS[1:3]
s <- scNMFSet(mat)
head(cells(s))
}
