% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{coeff,scNMFSet-method}
\alias{coeff,scNMFSet-method}
\title{Coefficient matrix accessor}
\usage{
\S4method{coeff}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing coefficient matrix}
}
\value{
List of coefficient matrices
}
\description{
Coefficient matrix accessor
}
