% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{coeff<-,scNMFSet-method}
\alias{coeff<-,scNMFSet-method}
\title{Modify coefficient matrices}
\usage{
\S4method{coeff}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Coefficient matrix to be substituted}
}
\value{
Input object with updated coefficient matrices
}
\description{
Can be used to access and modify coefficient matrices
}
\examples{
s <- scNMFSet(count=matrix(rpois(n=12,lambda=3),4,3))
s <- vb_factorize(s, ranks=3)
coeff(s)[[1]] <- apply(coeff(s)[[1]],1:2,round,digits=2)
coeff(s)
}
