% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{colData,scNMFSet-method}
\alias{colData,scNMFSet-method}
\title{Cell annotation accessor}
\usage{
\S4method{colData}{scNMFSet}(x)
}
\arguments{
\item{x}{Object containing cell annotation}
}
\description{
Cell annotation accessor
}
\examples{
s <- scNMFSet(count=matrix(rpois(n=12,lambda=3),4,3))
colData(s) <- letters[1:3]
colData(s)
}
