% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{colData<-,scNMFSet,ANY-method}
\alias{colData<-,scNMFSet,ANY-method}
\title{Cell annotation assignment}
\usage{
\S4method{colData}{scNMFSet,ANY}(x) <- value
}
\arguments{
\item{x}{Object containing cell annotation}

\item{value}{DataFrame to be substituted}
}
\description{
Cell annotation assignment
}
\examples{
s <- scNMFSet(count=matrix(rpois(n=12,lambda=3),4,3))
colData(s) <- letters[1:3]
colData(s)
}
