% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{count,scNMFSet-method}
\alias{count,scNMFSet-method}
\title{Count matrix accessor}
\usage{
\S4method{count}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing count matrix}
}
\value{
Count matrix
}
\description{
Count matrix accessor
}
