% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{count<-,scNMFSet-method}
\alias{count<-,scNMFSet-method}
\title{Count matrix assignment}
\usage{
\S4method{count}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing count matrix}

\item{value}{New count matrix to be substituted}
}
\value{
Object with count modified
}
\description{
Count matrix assignment
}
