% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{count<-}
\alias{count<-}
\title{Modify count matrix}
\usage{
count(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{count matrix to be substituted}
}
\value{
Input object with updated count matrix
}
\description{
Can be used to access and modify count matrix
}
\examples{
s <- scNMFSet(matrix(0, 4,3))
count(s) <- matrix(rpois(n=12,lambda=3),4,3)
count(s)
}
