% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{count}
\alias{count}
\title{Count matrix of Object}
\usage{
count(object)
}
\arguments{
\item{object}{Object of class \code{scNMFSet}.}
}
\value{
Either \code{NULL} or count matrix of class 
\code{dgCMatrix}.

Count matrix in sparse format
}
\description{
Retrieve or set the count matrix of an object
}
\details{
\code{count(object)} will return the count matrix. 
\code{count(object) <- value}
can be used to assign to the count slot of \code{object}.
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
head(count(s))
}
