% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{counts,scNMFSet-method}
\alias{counts,scNMFSet-method}
\title{Accessor for count matrix}
\usage{
\S4method{counts}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing count matrix}
}
\value{
Count matrix
}
\description{
Accessor for count matrix
}
\examples{
s <- scNMFSet(count = matrix(rpois(n=12,lambda=3),3,4))
counts(s)

}
