% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{counts<-,scNMFSet-method}
\alias{counts<-,scNMFSet-method}
\title{Assignment of count matrix}
\usage{
\S4method{counts}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing count}

\item{value}{Matrix-like object for replacement}
}
\value{
Object with updated count
}
\description{
Count matrix can be modified
}
\examples{
mat <- matrix(rpois(n=12,lambda=3),3,4)
s <- scNMFSet(count = mat)
counts(s) <- mat^2
counts(s)
}
