% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{dbasis,scNMFSet-method}
\alias{dbasis,scNMFSet-method}
\title{Basis SD matrix accessor}
\usage{
\S4method{dbasis}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing basis standard deviation (SD) matrix}
}
\value{
List of dbasis matrices
}
\description{
Basis SD matrix accessor
}
