% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{dbasis<-,scNMFSet-method}
\alias{dbasis<-,scNMFSet-method}
\title{Modify dbasis matrices}
\usage{
\S4method{dbasis}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Basis SD matrix to be substituted}
}
\value{
Modified object
}
\description{
Access and modify dbasis matrices
}
