% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{dbasis<-}
\alias{dbasis<-}
\title{Basis SD matrix assignment}
\usage{
dbasis(object) <- value
}
\arguments{
\item{object}{Object containing dbasis matrix}

\item{value}{List for assignment}
}
\value{
Updated object
}
\description{
Basis SD matrix assignment
}
