% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{dbasis}
\alias{dbasis}
\title{Basis SD matrix accessor}
\usage{
dbasis(object)
}
\arguments{
\item{object}{Object containing dbasis matrix}
}
\value{
List of dbasis matrices
}
\description{
Basis SD matrix accessor
}
