% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{dcoeff,scNMFSet-method}
\alias{dcoeff,scNMFSet-method}
\title{Coeffcient SD matrix accessor}
\usage{
\S4method{dcoeff}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing coeffient standard deviation (SD) matrix}
}
\value{
List of dcoeff matrices
}
\description{
Coeffcient SD matrix accessor
}
