% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{dcoeff<-,scNMFSet-method}
\alias{dcoeff<-,scNMFSet-method}
\title{Modify dcoeff matrices}
\usage{
\S4method{dcoeff}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Coeff SD matrix to be substituted}
}
\value{
Updated object
}
\description{
Access and modify dcoeff matrices
}
