% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{dcoeff<-}
\alias{dcoeff<-}
\title{Coeff SD matrix assignment}
\usage{
dcoeff(object) <- value
}
\arguments{
\item{object}{Object containing dcoeff matrix}

\item{value}{List for assignment}
}
\value{
Updated object
}
\description{
Coeff SD matrix assignment
}
