% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{dcoeff}
\alias{dcoeff}
\title{Coeff SD matrix accessor}
\usage{
dcoeff(object)
}
\arguments{
\item{object}{Object containing dcoeff matrix}
}
\value{
List of dcoeff matrices
}
\description{
Coeff SD matrix accessor
}
