% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{dim,scNMFSet-method}
\alias{dim,scNMFSet-method}
\title{Dimension of an object}
\usage{
\S4method{dim}{scNMFSet}(x)
}
\arguments{
\item{x}{Object of class \code{scNMFSet}.}
}
\value{
Vector of dimensions.
}
\description{
Retrieve dimension of an object.
}
\details{
\code{(dim(object))} will return dimesions of \code{object}
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
dim(s)
}
