% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{dmeasure,scNMFSet-method}
\alias{dmeasure,scNMFSet-method}
\title{Null rank measure accessor}
\usage{
\S4method{dmeasure}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing null measure}
}
\value{
Data frame of null measure
}
\description{
Null rank measure accessor
}
