% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{dmeasure<-,scNMFSet-method}
\alias{dmeasure<-,scNMFSet-method}
\title{Null rank measure assignment}
\usage{
\S4method{dmeasure}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing null measure}

\item{value}{New null measure (data frame) to be substituted}
}
\value{
Object with null measure modified
}
\description{
Null rank measure assignment
}
