% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{dmeasure<-}
\alias{dmeasure<-}
\title{Modify factorization null measures
 
Can be used to access and modify factorization null measures}
\usage{
dmeasure(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Null measure (data frame) to be substituted}
}
\value{
Input object with updated \code{dmeasure}

Input object with updated dmeasure
}
\description{
Modify factorization null measures
 
Can be used to access and modify factorization null measures
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
s <- factorize(s, ranks=2:4)
s <- factorize_H0(s)
dmeasure(s)[,-1] <- apply(dmeasure(s)[,-1],1:2, round,digits=3)
dmeasure(s)
}
