% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{dmeasure}
\alias{dmeasure}
\title{Quality measures under null hypotheses}
\usage{
dmeasure(object)
}
\arguments{
\item{object}{Object of class \code{scNMFSet}.}
}
\value{
Either \code{NULL} or a data frame containing difference measures.
}
\description{
Retrieve or set net quality measures in an object
}
\details{
Factorization using maximum likelihood using \code{\link{factorize}} 
can be followed 
by \code{\link{factorize_H0}}, which repeats inference for all rank 
values with
randomly permuted data. The differences in quality measures after 
subtracting
the null values are stored in \code{dmeasure} slot. 
\code{dmeasure(object)} 
will return the data frame. \code{dmeasure(object) <- value}
can be used to modify it. Not used when factorization is done with 
\code{\link{vb_factorize}}.
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
s <- factorize(s,ranks=2:4)
s <- factorize_H0(s)
dmeasure(s)
}
