% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize_H0}
\alias{factorize_H0}
\title{Maximum likelihood factorization with randomization}
\usage{
factorize_H0(object, ...)
}
\arguments{
\item{object}{\code{scNMFSet} object containing 
count matrix and factorization results}

\item{...}{Other parameters to \code{factorize()}.}
}
\value{
Input \code{object} with slot \code{dmeasure} containing 
       differences in quality measures between those for original data
       and H0 results.
}
\description{
Updates \code{scNMFSet object} by performing factorizations with
shuffled data matrices.
}
\details{
This function updates the main object of class \code{scNMFSet} 
         by performing factorizations of randomized count data. It is assumed
         that main factorization of the original data has already been 
         performed. The same set of rank values in \code{ranks} slot of 
         \code{object} is used for factorization. The function 
         \code{factorize} 
         is then called with appropriate parameters. The resulting quality
         measures are stored in the slot
         \code{dmeasure}, a data frame with columns \code{likelihood} 
         and \code{dispersion}.
}
\examples{
set.seed(1)
x <- simulate_data(nfeatures=10,nsamples=c(20,20,60,40,30))
s <- scNMFSet(x)
s <- factorize(s,ranks=2:8)
s <- factorize_H0(s)
plot(s)
}
