% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{genes,scNMFSet-method}
\alias{genes,scNMFSet-method}
\title{Gene annotation accessor}
\usage{
\S4method{genes}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing gene annotation}
}
\value{
Gene annotation data frame
}
\description{
Gene annotation accessor
}
