% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{genes<-,scNMFSet-method}
\alias{genes<-,scNMFSet-method}
\title{Gene annotation assignment}
\usage{
\S4method{genes}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing gene annotation}

\item{value}{New gene annotation to be substituted}
}
\value{
Object with modified gene annotation
}
\description{
Gene annotation assignment
}
