% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{genes<-}
\alias{genes<-}
\title{Modify gene annotation
 
Can be used to access and modify gene annotation}
\usage{
genes(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Gene annotation data frame to be substituted}
}
\value{
Input object with updated gene annotation

Input object with updated gene annotation
}
\description{
Modify gene annotation
 
Can be used to access and modify gene annotation
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
genes(s) <- letters[1:3]
genes(s)
}
