% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{genes}
\alias{genes}
\title{Feature set of Object}
\usage{
genes(object)
}
\arguments{
\item{object}{Object of class \code{scNMFSet}.}
}
\value{
Either \code{NULL} or data frame.
}
\description{
Retrieve or set the gene annotation of object
}
\details{
\code{genes(object)} will return the gene annotation. 
\code{genes(object) <- value}
can be used to assign to the \code{genes} slot of \code{object}.
}
\examples{
mat <- matrix(rpois(n=12,lambda=3),4,3)
rownames(mat) <- letters[1:4]
s <- scNMFSet(mat)
head(genes(s))
}
