% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{measure,scNMFSet-method}
\alias{measure,scNMFSet-method}
\title{Rank measure accessor}
\usage{
\S4method{measure}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing measure}
}
\value{
Data frame of measure
}
\description{
Rank measure accessor
}
