% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{measure<-,scNMFSet-method}
\alias{measure<-,scNMFSet-method}
\title{Rank measure assignment}
\usage{
\S4method{measure}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing measure}

\item{value}{New measure (data frame) to be substituted}
}
\value{
Object with modified measure
}
\description{
Rank measure assignment
}
