% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{ranks,scNMFSet-method}
\alias{ranks,scNMFSet-method}
\title{Rank accessor}
\usage{
\S4method{ranks}{scNMFSet}(object)
}
\arguments{
\item{object}{Object containing rank values}
}
\value{
Vector of rank values
}
\description{
Rank accessor
}
