% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{ranks<-,scNMFSet-method}
\alias{ranks<-,scNMFSet-method}
\title{Rank values assignment}
\usage{
\S4method{ranks}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing ranks}

\item{value}{New rank vector to be substituted}
}
\value{
Object with rank vector modified
}
\description{
Rank values assignment
}
