% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{ranks<-}
\alias{ranks<-}
\title{Modify ranks
 
Can be used to access and modify ranks}
\usage{
ranks(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Rank values (vector) to be substituted}
}
\value{
Input object with updated ranks

Input object with updated ranks
}
\description{
Modify ranks
 
Can be used to access and modify ranks
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
s <- vb_factorize(s, ranks=2:3)
ranks(s) <- c('two','three')
ranks(s)
}
