% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{remove_zeros}
\alias{remove_zeros}
\title{Remove rows or columns that are empty from an object}
\usage{
remove_zeros(object)
}
\arguments{
\item{object}{Object containing data}
}
\value{
Object with empty rows/columns removed
}
\description{
Remove rows or columns that are empty from an object
}
\examples{
set.seed(1)
x <- matrix(rpois(n=100,lambda=0.1),10,10)
s <- scNMFSet(count=x,remove.zeros=FALSE)
s2 <- remove_zeros(s)
s2
}
