% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{rowData<-,scNMFSet-method}
\alias{rowData<-,scNMFSet-method}
\title{Gene annotation assignment}
\usage{
\S4method{rowData}{scNMFSet}(x) <- value
}
\arguments{
\item{x}{Object containing data}

\item{value}{DataFrame of row annotation to be substituted}
}
\description{
Gene annotation assignment
}
