% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{class}
\name{scNMFSet-class}
\alias{scNMFSet-class}
\alias{scNMFSet}
\alias{plot,scNMFSet-method}
\title{Class for storing input data and results}
\usage{
\S4method{plot}{scNMFSet}(x, measure = "measure")
}
\arguments{
\item{x}{Object containing measure}

\item{measure}{\code{c('measure','dmeasure')}}
}
\value{
Object of class \code{scNMFSet}

\code{NULL}
}
\description{
\code{S4} class that can store single-cell count matrix, 
gene and cell 
annotation data frames, and factorization factors as well 
as quality measures
for rank determination.
}
\section{Methods (by generic)}{
\itemize{
\item \code{plot}: Plot measures of an object. For quality 
measures derived from
maximum likelihood inference, dispersion and cophenetic will 
be plotted separately.

For measure derived from Bayesian inference, log evidence as a function of 
rank values will be plotted. Parameter 
\code{measure = c('measure','dmeasure')} can be used
when slot \code{dmeasure} has been filled (maximum likelihood NMF)
to plot either H1 or H0 quality measures.
}}

\section{Slots}{

\describe{
\item{\code{count}}{Matrix containing count data.}

\item{\code{genes}}{Data frame with row names containing gene (feature) 
names and annotations in columns.}

\item{\code{cells}}{Data frame with cell IDs in row names and other 
annotations in columns (e.g., barcodes, cell types).}

\item{\code{ranks}}{Vector for rank values for which factorization has been 
performed.}

\item{\code{basis}}{List (of length equal to that of \code{ranks}) of 
basis matrices \strong{W} from factorization; 
dimension \code{nrow} x \code{rank},
    where \code{nrow} is no. of rows in \code{count}.}

\item{\code{coeff}}{List (of length equal to that of \code{ranks}) of 
coefficient matrices \strong{H} from factorization; 
     dimension \code{rank} x \code{ncol},
     where \code{ncol} is no. of columns in \code{count}.}

\item{\code{measure}}{Data frame of factorization quality measures for 
each rank (\code{likelihood} and \code{dispersion}).}

\item{\code{dmeasure}}{Data frame of quality measures with randomized
matrix counts (H0).}
}}

\examples{
# toy matrix
ngenes <- 8; ncells <- 5
mat <- matrix(rpois(n=ngenes*ncells,lambda=3),ngenes,ncells)

abc <- letters[1:ngenes]
ABC <- LETTERS[1:ncells] 
genes <- data.frame(gene_id=abc)
cells <- data.frame(cell_id=ABC)
rownames(mat) <- rownames(genes) <- abc
colnames(mat) <- rownames(cells) <- ABC

# create scNMFSet object
s <- scNMFSet(count=mat,genes=genes,cells=cells)
# alternative ways (genes and cells slots will be filled from count)
s <- scNMFSet(count=mat)
s <- scNMFSet(mat)        

# show dimensions
s

# show slots
head(genes(s))

# modify slots
cells(s) <- data.frame(cell_id=1:ncells,
            cell_type=c(rep('tissue1',2),
            rep('tissue2',ncells-2)))
# slots can be accessed directly
head(cells(s))  

}
