% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{show,scNMFSet-method}
\alias{show,scNMFSet-method}
\title{Display object}
\usage{
\S4method{show}{scNMFSet}(object)
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}
}
\value{
\code{NULL}
}
\description{
Display the class and dimension of an object
 
Object name itself on command line or \code{(show(object))} will display 
class and dimensionality
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
show(s)
}
