% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{[,scNMFSet,ANY,ANY,ANY-method}
\alias{[,scNMFSet,ANY,ANY,ANY-method}
\alias{[,scNMFSet-method}
\title{Subsetting scNMFSet object}
\usage{
\S4method{[}{scNMFSet,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{x}{Object to be subsetted}

\item{i}{row index}

\item{j}{column index}
}
\value{
Subsetted object
}
\description{
Subsetting scNMFSet object
}
