% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{coeff<-,scNMFSet-method}
\alias{coeff<-,scNMFSet-method}
\title{Coefficient matrix assignment}
\usage{
\S4method{coeff}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object containing coefficient matrices}

\item{value}{New coefficient matrices to be substituted}
}
\value{
Object with coefficient matrix modified
}
\description{
Coefficient matrix assignment
}
