% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{coeff<-}
\alias{coeff<-}
\title{Modify coefficient matrices
 
Can be used to access and modify coefficient matrices}
\usage{
coeff(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Coefficient matrix to be substituted}
}
\value{
Input object with updated coefficient matrices

Input object with updated coefficient matrices
}
\description{
Modify coefficient matrices
 
Can be used to access and modify coefficient matrices
}
\examples{
s <- scNMFSet(matrix(rpois(n=12,lambda=3),4,3))
s <- vb_factorize(s, ranks=3)
coeff(s)[[1]] <- apply(coeff(s)[[1]],1:2,round,digits=2)
coeff(s)
}
