% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{counts,scNMFSet-method}
\alias{counts,scNMFSet-method}
\title{Accessor for count matrix}
\usage{
\S4method{counts}{scNMFSet}(object)
}
\arguments{
\item{Object}{Object containing count matrix}
}
\description{
Accessor for count matrix
}
\examples{
s <- scNMFSet(count = matrix(rpoise(n=12, 3,4)))
counts(s)

}
