% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{feature_map}
\alias{feature_map}
\title{Plot heatmap of basis matrix}
\usage{
feature_map(object, basis.matrix = NULL, rank, markers = NULL,
  subtract.mean = TRUE, log = TRUE, scheme = "max", sweep = TRUE,
  ntop = 1, max.per.cluster = 10, Colv = NA, feature.names = NULL,
  perm = NULL, main = "Feature map", cscale = NULL, cex.cluster = 1,
  cex.feature = 0.5, mar = NULL, ...)
}
\description{
Generate heatmap of features derived from factorization of count data.
#' @examples  
set.seed(1)
x <- simulate_data(nfeatures=10,nsamples=c(20,20,60))
rownames(x) <- seq_len(10)
colnames(x) <- seq_len(100)
s <- scNMFSet(count=x,rowData=seq_len(10), colData=seq_len(100))
s <- vb_factorize(s,ranks=seq(2,5))
plot(s)
feature_map(s, rank=3)
}
