% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{ranks<-,scNMFSet-method}
\alias{ranks<-,scNMFSet-method}
\title{Modify ranks}
\usage{
\S4method{ranks}{scNMFSet}(object) <- value
}
\arguments{
\item{object}{Object of class \code{scNMFSet}}

\item{value}{Rank values (vector) to be substituted}
}
\value{
Input object with updated ranks
}
\description{
Replace \code{ranks} slot of \code{scNMFSet} object
}
\examples{
s <- scNMFSet(count=matrix(rpois(n=12,lambda=3),4,3))
s <- vb_factorize(s, ranks=seq(2,3))
ranks(s) <- c('two','three')
ranks(s)
}
