% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\docType{methods}
\name{rowData,scNMFSet-method}
\alias{rowData,scNMFSet-method}
\title{Gene annotation accessor}
\usage{
\S4method{rowData}{scNMFSet}(x)
}
\arguments{
\item{x}{Object containing data}
}
\value{
DataFrame of row annotation
}
\description{
Gene annotation accessor
}
\examples{
x <- matrix(rpois(n=12,lambda=3),4,3)
rownames(x) <- 1:4
colnames(x) <- 1:3
s <- scNMFSet(count=x,rowData=1:4,colData=1:3)
rowData(s) 
}
