% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scNMF_class.R
\name{scNMFSet}
\alias{scNMFSet}
\title{Create \code{scNMFSet} object}
\usage{
scNMFSet(count = NULL, ..., remove.zeros = TRUE)
}
\arguments{
\item{count}{Count matrix}

\item{...}{Other parameters of \code{\link{SingleCellExperiment}}}

\item{remove.zeros}{Remove empty rows and columns}
}
\value{
Object of class \code{scNMFSet}.
}
\description{
Object derived from \code{\link{SingleCellExperiment}}
}
\examples{
count <- matrix(rpois(n=12,lambda=2),4,3)
s <- scNMFSet(count=count)
s
}
